+++ HLSpriteKit/HLMenuNode.h
@@ -7,6 +7,7 @@
 //
 
 #import <SpriteKit/SpriteKit.h>
+#import <TargetConditionals.h>
 
 #import "HLComponentNode.h"
 #import "HLGestureTarget.h"
@@ -43,10 +44,22 @@ typedef NS_ENUM(NSInteger, HLMenuNodeAnimation) {
  item in the `HLMenu`, stacked vertically.  Menus are hierarchical in nature, and the
  node provides functionality for navigating between menus and submenus.
 
- ## Common Gesture Handling Configurations
+ ## Common User Interaction Configurations
+
+ As a gesture target:
 
  - Set this node as its own gesture target (via `[SKNode+HLGestureTarget
    hlSetGestureTarget]`) to get a callbacks via the `HLMenuNodeDelegate` interface.
+
+ As a `UIResponder`:
+
+ - Set this node's `userInteractionEnabled` property to true to get callbacks via the
+  `HLMenuNodeDelegate` interface.
+
+ As an `NSResponder`:
+
+ - Set this node's `userInteractionEnabled` property to true to get callbacks via the
+  `HLMenuNodeDelegate` interface.
 */
 #if HLGESTURETARGET_AVAILABLE
 @interface HLMenuNode : HLComponentNode <NSCoding, HLGestureTarget>
@@ -194,20 +207,21 @@ typedef NS_ENUM(NSInteger, HLMenuNodeAnimation) {
 /**
  The animation used for navigation between menus.
 
- This animation applies when navigation is triggered by taps on submenu and back buttons.
+ This animation applies when navigation is triggered by taps or clicks on submenu and back
+ buttons.
 */
 @property (nonatomic, assign) HLMenuNodeAnimation itemAnimation;
 
 /**
  The duration of the animation used for navigation between menus.
 
- This animation duration applies when navigation is triggered by taps on submenu and back
- buttons.
+ This animation duration applies when navigation is triggered by taps or clicks on submenu
+ and back buttons.
 */
 @property (nonatomic, assign) NSTimeInterval itemAnimationDuration;
 
 /**
- Default sound file for playing when an item is tapped, or `nil` for none.
+ Default sound file for playing when an item is tapped or clicked, or `nil` for none.
 
  This sound file is used if no other more-specific sound file is specified.  In
  particular, a sound file for an item is found in the following order, from most-specific
@@ -239,10 +253,6 @@ typedef NS_ENUM(NSInteger, HLMenuNodeAnimation) {
 
 /**
  A delegate for `HLMenuNode`.
-
- The delegate is (currently) concerned mostly with handling user interaction.  It's
- worth noting that the `HLMenuNode` only receives gestures if it is configured as its
- own gesture target (using `[SKNode+HLGestureTarget hlSetGestureTarget]`).
 */
 @protocol HLMenuNodeDelegate <NSObject>
 
@@ -260,6 +270,8 @@ typedef NS_ENUM(NSInteger, HLMenuNodeAnimation) {
 
 /// @name Handling User Interaction
 
+#if TARGET_OS_IPHONE
+
 /**
  Called when the user taps on a menu item, but before the menu node has taken any
  navigation actions that would normally result from the tap.
@@ -268,6 +280,9 @@ typedef NS_ENUM(NSInteger, HLMenuNodeAnimation) {
 
  The sound file associated with the item will play **before** this delegate method is
  called.
+
+ Relevant to `HLGestureTarget` and `UIResponder` user interaction.
+ See "Common User Interaction Configurations".
 */
 @optional
 - (BOOL)menuNode:(HLMenuNode *)menuNode shouldTapMenuItem:(HLMenuItem *)menuItem itemIndex:(NSUInteger)itemIndex;
@@ -275,15 +290,59 @@ typedef NS_ENUM(NSInteger, HLMenuNodeAnimation) {
 /**
  Called when the user has tapped on a menu item and the menu node has taken any
  navigation actions that would normally result from the tap.
+
+ Relevant to `HLGestureTarget` and `UIResponder` user interaction.
+ See "Common User Interaction Configurations".
 */
 @required
 - (void)menuNode:(HLMenuNode *)menuNode didTapMenuItem:(HLMenuItem *)menuItem itemIndex:(NSUInteger)itemIndex;
 
 /**
  Called when the user has long-pressed on a menu item.
+
+ Relevant to `HLGestureTarget` and `UIResponder` user interaction.
+ See "Common User Interaction Configurations".
 */
 @optional
 - (void)menuNode:(HLMenuNode *)menuNode didLongPressMenuItem:(HLMenuItem *)menuItem itemIndex:(NSUInteger)itemIndex;
+
+#else
+
+/**
+ Called when the user clicks on a menu item, but before the menu node has taken any
+ navigation actions that would normally result from the click.
+
+ A return value of `YES` indicates the menu node should continue to navigation.
+
+ The sound file associated with the item will play **before** this delegate method is
+ called.
+
+ Relevant to `NSResponder` user interaction.
+ See "Common User Interaction Configurations".
+*/
+@optional
+- (BOOL)menuNode:(HLMenuNode *)menuNode shouldClickMenuItem:(HLMenuItem *)menuItem itemIndex:(NSUInteger)itemIndex;
+
+/**
+ Called when the user has clicked on a menu item and the menu node has taken any
+ navigation actions that would normally result from the click.
+
+ Relevant to `NSResponder` user interaction.
+ See "Common User Interaction Configurations".
+*/
+@required
+- (void)menuNode:(HLMenuNode *)menuNode didClickMenuItem:(HLMenuItem *)menuItem itemIndex:(NSUInteger)itemIndex;
+
+/**
+ Called when the user has long-clicked on a menu item.
+
+ Relevant to `NSResponder` user interaction.
+ See "Common User Interaction Configurations".
+ */
+@optional
+- (void)menuNode:(HLMenuNode *)menuNode didLongClickMenuItem:(HLMenuItem *)menuItem itemIndex:(NSUInteger)itemIndex;
+
+#endif
 
 @end
 
@@ -323,7 +382,7 @@ typedef NS_ENUM(NSInteger, HLMenuNodeAnimation) {
 @property (nonatomic, strong) SKNode *buttonPrototype;
 
 /**
- The sound file that will be played when this button is tapped, if not `nil`.
+ The sound file that will be played when this button is tapped or clicked, if not `nil`.
 
  See notes in `[HLMenuNode itemSoundFile]`.
 */
@@ -419,8 +478,8 @@ typedef NS_ENUM(NSInteger, HLMenuNodeAnimation) {
 @end
 
 /**
- An HLMenuBackItem is a kind of menu item which, when tapped, navigates the menu node to
- its parent menu.
+ An HLMenuBackItem is a kind of menu item which, when tapped or clicked, navigates the
+ menu node to its parent menu.
 */
 @interface HLMenuBackItem : HLMenuItem <NSCoding>